sample_cdr_georef/pyproject.toml
[tool.poetry]
name = "sample-cdr-georef"
version = "0.0.1"
description = ""
authors = []
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
uvicorn = "^0.27.1"
fastapi = "^0.110.0"
pydantic = "^2.6.3"
httpx = "^0.27.0"
pillow = "^10.1.0"
pydantic-settings = "^2.2.1"
pyproj = "^3.6.1"
rasterio = "^1.3.9"
ngrok = "^1.2.0"
cdr_schemas = { git = "https://github.com/DARPA-CRITICALMAAS/cdr_schemas.git", tag = "v0.1.2+draft" }


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
black = "^24.2.0"
autoflake = "^2.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

